navigator: move single-concern controller files to stores/ and util/
Step 0 of the controllers/ refactor (see docs/refactor-controllers.md):
hoist files that already have a single concern into the new top-level
layout. Multi-concern files (app.ts, controls.ts, nav-sheet.ts) stay put
until later steps split them.

- controllers/ui-environment.ts -> stores/ui-environment.ts
- controllers/map-padding.ts    -> stores/map-padding.ts
- controllers/telemetry-timeline.ts -> util/telemetry-timeline.ts

No logic changes. Imports updated in the moved files plus create-app.tsx
and controllers/app.ts.
