+++ packages/apps/navigator/src/controllers/app.ts
@@ -23,8 +23,8 @@ import { action, makeAutoObservable, observable, runInAction } from 'mobx';
 import type { MapRef } from 'react-map-gl/maplibre';
 import { lineGradientExpression } from '../components/RoutesStyle';
 import { toRouteFeatures } from '../route-features';
+import { TelemetryTimeline } from '../util/telemetry-timeline';
 import { BearingMode, CameraMode } from './constants';
-import { TelemetryTimeline } from './telemetry-timeline';
 import type { AppClient, AppController, AppStore } from './types';
 
 export class AppStoreImpl implements AppStore {
