+++ packages/apps/navigator/src/util/telemetry-timeline.ts
